travel class minor changes

+++ src/main/java/com/dmytrobozhor/airlinereservationservice/service/AbstractTravelClassService.java
@@ -4,6 +4,6 @@ import com.dmytrobozhor.airlinereservationservice.domain.TravelClass;
 
 import java.util.Optional;
 
-public interface AbstractTravelClassService extends AbstractCrudService<TravelClass, Integer> {
+public interface AbstractTravelClassService extends AbstractCrudService<TravelClass, Long> {
 
 }
